# server/docs/swagger.yaml
definitions:
...
  CreateMemberRequest:
    properties:
      email:
        type: string
      faculty:
        enum:
        - AHS
        - Arts
        - Engineering
        - Environment
        - Math
        - Science
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      questId:
        type: string
    required:
    - email
    - firstName
    - id
    - lastName
    type: object
...
  ErrorResponse:
    properties:
      code:
        description: Code is the HTTP status code for the error.
        type: integer
      message:
        description: Message is a more descriptive error message.
        type: string
      type:
        description: Type is the name of the HTTP status code.
        type: string
    type: object
...
  Member:
    properties:
      createdAt:
        type: string
      email:
        type: string
      faculty:
        enum:
        - AHS
        - Arts
        - Engineering
        - Environment
        - Math
        - Science
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      questId:
        type: string
    required:
    - email
    - firstName
    - id
    - lastName
    type: object
...
  UpdateMemberRequest:
    properties:
      email:
        type: string
      faculty:
        enum:
        - AHS
        - Arts
        - Engineering
        - Environment
        - Math
        - Science
        type: string
      firstName:
        type: string
      lastName:
        type: string
      questId:
        type: string
    type: object
...
info:
  contact:
    email: uwaterloopoker@gmail.com
    name: UWPSC Development Team
  description: This is the API for the UWPSC website.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: UWPSC API
  version: "1.0"
paths:
...
  /members:
    get:
      consumes:
      - application/json
      description: Retrieve a list of Members with optional filters
      parameters:
      - description: Filter by Member ID
        in: query
        name: id
        type: integer
      - description: Filter by Member Email
        in: query
        name: email
        type: string
      - description: Filter by Member Name
        in: query
        name: name
        type: string
      - description: Filter by Member Faculty
        in: query
        name: faculty
        type: string
      - description: Filter by exact Quest ID
        in: query
        name: questId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Member'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: List Members
      tags:
      - Members
    post:
      consumes:
      - application/json
      description: Create a new Member with the provided details
      parameters:
      - description: Member details
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/CreateMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Member'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create a new Member
      tags:
      - Members
  /members/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Member by their ID
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Delete Member by ID
      tags:
      - Members
    get:
      consumes:
      - application/json
      description: Retrieve a Member by their ID
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Member'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get Member by ID
      tags:
      - Members
    patch:
      consumes:
      - application/json
      description: Update a Member's details by their ID
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Member details
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/UpdateMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Member'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Update Member by ID
      tags:
      - Members
...
swagger: "2.0"